+++ .gitignore
@@ -4,9 +4,9 @@ examples/js/clingo.js
 examples/js/clingo.wasm
 lua-*
 node_modules
+wallaby.js
 # Build products
-node.js
-web.js
+dist/**/*
 # Webpack gives the WASM file a hashed name.
 *.wasm
 # Unignore the source file.
