--- wallaby.js
@@ -1,16 +1,0 @@
-module.exports = function () {
-    return {
-        files: [
-            'src/**/*.ts',
-            'index.node.ts',
-            'clingo.js',
-            '*.wasm',
-        ],
-        tests: [
-            'test/**/*.ts',
-        ],
-        env: {
-            type: 'node'
-        }
-    };
-};
